requirements.txt
Flask==1.1.2
pytest==6.1.2
pytest-flask==1.1.0
python-dotenv==0.15.0